refactor: moved home page components in it's own feature folder
