+++ src/features/store/homePage/components/categories/catListItem.tsx
