+++ src/features/store/homePage/components/categories/index.tsx
