+++ src/features/store/homePage/components/companyLogo/CompanyLogo.tsx
