+++ src/features/store/homePage/components/companyLogo/index.tsx
