+++ src/features/store/homePage/components/slider/index.tsx
