+++ src/features/store/homePage/components/todayDealCard/index.tsx
