+++ src/features/store/homePage/components/wideAd/index.tsx
